task test

+++ server/src/main/java/ru/trandefil/tm/service/TaskServiceImpl.java
@@ -4,9 +4,10 @@ import ru.trandefil.tm.api.TaskRepository;
 import ru.trandefil.tm.api.TaskService;
 import ru.trandefil.tm.entity.Session;
 import ru.trandefil.tm.entity.Task;
+import ru.trandefil.tm.util.SignatureUtil;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class TaskServiceImpl implements TaskService {
 
@@ -17,22 +18,44 @@ public class TaskServiceImpl implements TaskService {
     }
 
     public List<Task> getAll(Session session) {
-        return new ArrayList(taskRepository.getAll());
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
+        return taskRepository.getAll().stream()
+                .filter(t -> t.getExecuterId().equals(session.getUserId()))
+                .collect(Collectors.toList());
     }
 
     public Task save(Task task, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return taskRepository.save(task);
     }
 
     public Task delete(Task task, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return taskRepository.delete(task);
     }
 
     public Task deleteByName(String name, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return taskRepository.deleteByName(name);
     }
 
     public Task getByName(String name, Session session) {
+        if (!SignatureUtil.checkCorrectSession(session)) {
+            System.out.println("bad signature.");
+            return null;
+        }
         return taskRepository.getByName(name);
     }
